Moving map and map_df_uses_f to tensor_ops/utils. Making them pub(crate)
